ref: fixing comments
